refactor: 모델 컨트롤러에 있던 트림 목록 조회 API를 트림 컨트롤러로 이전
